+++ packages/ui/src/memberships/components/ProfileComponent.tsx
@@ -1,12 +1,13 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { QuestionIcon } from '@/common/components/icons'
+import { QuestionIcon, WatchIcon } from '@/common/components/icons'
 import { Tooltip, DefaultTooltip } from '@/common/components/Tooltip'
+import { useLocalStorage } from '@/common/hooks/useLocalStorage'
 
 import { TransferButtonStyled } from '../../accounts/components/TransferButton'
 import { useMyTotalBalances } from '../../accounts/hooks/useMyTotalBalances'
-import { TextSmall, TokenValue } from '../../common/components/typography'
+import { TextInlineMedium, TextSmall, TokenValue } from '../../common/components/typography'
 import { Colors } from '../../common/constants'
 
 import { Memberships } from '.'
@@ -14,15 +15,21 @@ import { CurrentMember } from './CurrentMember'
 
 export function ProfileComponent() {
   const { total } = useMyTotalBalances()
+  const [isBalanceHidden, setIsBalanceHidden] = useLocalStorage<boolean | undefined>('isBalanceHidden')
 
   return (
     <Profile>
       <CurrentMember />
       <MemberBalance>
-        <BalanceTitle>Total Balance</BalanceTitle>
+        <BalanceTitle>
+          Total Balance
+          <StyledWatchIconContainer onClick={() => setIsBalanceHidden(!isBalanceHidden)}>
+            <StyledWatchIcon />
+          </StyledWatchIconContainer>
+        </BalanceTitle>
         <TotalBalance>
           <TokenDetail>
-            <TotalTokenValue value={total} />
+            {isBalanceHidden ? <TextInlineMedium>*****</TextInlineMedium> : <TotalTokenValue value={total} />}
             <StyledTooltip
               tooltipLinkText="Learn how to earn JOY’s"
               tooltipLinkURL="https://www.joystream.org/token#earn"
@@ -80,6 +87,8 @@ const MemberBalance = styled.div`
 `
 
 export const BalanceTitle = styled.span`
+  display: flex;
+  column-gap: 16px;
   grid-area: balancetitle;
   font-size: 10px;
   line-height: 16px;
@@ -127,4 +136,16 @@ const StyledTooltip = styled(Tooltip)`
     background-color: ${Colors.Black[75]};
     border-color: ${Colors.Black[300]};
   }
+`
+const StyledWatchIconContainer = styled.div`
+  display: flex;
+  position: relative;
+  justify-content: center;
+  align-items: center;
+  width: 16px;
+  height: 16px;
+  cursor: pointer;
+`
+const StyledWatchIcon = styled(WatchIcon)`
+  marign-top: 4px;
 `