+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2024-01-18
+
+### Added
+- Display blacklisted accounts in the election module.
+- Allow hiding the sidebar "Total balance".
+
+### Changed
+- Link block heights to the Subscan block explorer.
+
+### Fixed
+- Fix past election page endless refresh issue.
+
+## [2.3.1] - 2024-01-12
+
+### Fixed
+- On boarding modal footer.
+
 ## [2.3.0] - 2023-12-29
 
 ### Added
@@ -270,7 +287,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2022-12-02
 
-[unreleased]: https://github.com/Joystream/pioneer/compare/v2.3.0...HEAD
+[unreleased]: https://github.com/Joystream/pioneer/compare/v2.4.0...HEAD
+[2.4.0]: https://github.com/Joystream/pioneer/compare/v2.3.1...v2.4.0
+[2.3.1]: https://github.com/Joystream/pioneer/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/Joystream/pioneer/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/Joystream/pioneer/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/Joystream/pioneer/compare/v2.1.0...v2.1.1